feat(select,tree-select): moved common logic to separate class (#541)
