+++ tools/public_api_guard/components/core.api.md
@@ -11,6 +11,7 @@ import { AfterViewInit } from '@angular/core';
 import { AnimationEvent as AnimationEvent_2 } from '@angular/animations';
 import { AnimationTriggerMetadata } from '@angular/animations';
 import { BehaviorSubject } from 'rxjs';
+import { CdkConnectedOverlay } from '@angular/cdk/overlay';
 import { ChangeDetectorRef } from '@angular/core';
 import { ComponentPortal } from '@angular/cdk/portal';
 import { ConnectedOverlayPositionChange } from '@angular/cdk/overlay';
@@ -1186,6 +1187,23 @@ export const KBQ_TITLE_TEXT_REF: InjectionToken<KbqTitleTextRef>;
 
 // @public @deprecated (undocumented)
 export const KBQ_VALIDATION: InjectionToken<KbqValidationOptions>;
+
+// @public
+export abstract class KbqAbstractSelect {
+    // (undocumented)
+    protected calculateOverlayOffsetX(baseOffsetX: number): number[];
+    // (undocumented)
+    protected getOverlayRect(): DOMRect;
+    // (undocumented)
+    protected overlayDir: CdkConnectedOverlay;
+    protected readonly overlayPanelClass = "kbq-select-overlay";
+    // (undocumented)
+    protected resetOverlay(): void;
+    // (undocumented)
+    protected setOverlayPosition(): void;
+    // (undocumented)
+    protected triggerRect: DOMRect;
+}
 
 // @public
 export type KbqCodeBlockLocaleConfiguration = {
